Add: Maze Project is finished
Done with project 3. This is a VR maze that you have to use clickable waypoints to move through. It's cause it only supports 3DOF and doesn't have a controller implemented. It's kind of cool except for when it takes forever to get the focus of a waypoint. This is the final project for Term 1 of Udacity's VR nanodegree.